setup script で config.php の存在確認を追加

+++ setup.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
+if [ ! -e config.php ]; then
 cp config.php.sample config.php
+fi
 git submodule init
 git submodule foreach 'git co 1.7/master'
 git submodule update
